+++ R/misc.R
@@ -251,3 +251,18 @@ ReassignmentQP <- function(spectra, exposure, sigs, mc.cores = 1) {
   exp.all <- MergeListOfExposures(exp.list)
   return(exp.all)
 }
+
+GetSigActivity <- function(spectra, exposure, sigs, sig.id, output.dir) {
+  exposure.one.sig <- exposure[sig.id, ]
+  non.zero.exposure.samples <- names(exposure.one.sig[exposure.one.sig > 0])
+  spectra.to.use <- spectra[, non.zero.exposure.samples, drop = FALSE]
+  exposure.to.use <- exposure[, non.zero.exposure.samples, drop = FALSE]
+  
+  sig.activity <- AddSigActivity(spectra = spectra.to.use, 
+                                 exposure = exposure.to.use, 
+                                 sigs = sigs, 
+                                 use.sparse.assign = TRUE)
+  ShowSigActivity(list.of.sig.activity = sig.activity, 
+                  output.dir = output.dir)
+  return(RemoveZeroActivitySig(exposure.to.use))
+}